% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errno.R
\name{errno}
\alias{errno}
\title{List of 'errno' error codes}
\usage{
errno()
}
\description{
For the errors that are not used on the current platform, \code{value} is
\code{NA_integer_}.
}
\details{
A data frame with columns: \code{name}, \code{value}, \code{description}.
}
\examples{
\dontshow{if (ps::ps_is_supported() && ! ps:::is_cran_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
errno()
\dontshow{\}) # examplesIf}
}
