% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Create a system process object}
\usage{
process(pid = NULL)
}
\arguments{
\item{pid}{Process id, integer scalar. If not specified or \code{NULL},
the pid of the  calling process is used.}
}
\value{
\code{process} object.
}
\description{
TODO
}
