% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps-package.R
\docType{package}
\name{ps-package}
\alias{ps-package}
\alias{_PACKAGE}
\title{ps: List, Query, Manipulate System Processes}
\description{
List, query and manipulate all system processes, on 'Windows', 'Linux' and 'macOS'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/ps}
  \item \url{https://ps.r-lib.org/}
  \item Report bugs at \url{https://github.com/r-lib/ps/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Jay Loden
  \item Dave Daeschler
  \item Giampaolo Rodola'
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
