% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-level.R, R/system.R
\name{ps_cpu_times}
\alias{ps_cpu_times}
\title{CPU times of the process}
\usage{
ps_cpu_times()

ps_cpu_times()
}
\arguments{
\item{p}{Process handle.}
}
\value{
Named real vector or length four: \code{user}, \code{system},
\code{children_user},  \code{children_system}. The last two are \code{NA} on
non-Linux systems.

Named list
}
\description{
All times are measured in seconds:
\itemize{
\item \code{user}: Amount of time that this process has been scheduled in user
mode.
\item \code{system}: Amount of time that this process has been scheduled in
kernel mode
\item \code{children_user}: On Linux, amount of time that this process's
waited-for children have been scheduled in user mode.
\item \code{children_system}: On Linux, Amount of time that this process's
waited-for children have been scheduled in kernel mode.
}

Every attribute represents the seconds the CPU has spent in the given
mode. The attributes availability varies depending on the platform:
\itemize{
\item \code{user}: time spent by normal processes executing in user mode;
on Linux this also includes guest time.
\item \code{system}: time spent by processes executing in kernel mode.
\item \code{idle}: time spent doing nothing.
}
}
\details{
Throws a \code{zombie_process()} error for zombie processes.

Platform-specific fields:
\itemize{
\item \code{nice} (UNIX): time spent by niced (prioritized) processes executing
in user mode; on Linux this also includes guest_nice time.
\item \code{iowait} (Linux): time spent waiting for I/O to complete. This is not
accounted in idle time counter.
\item \code{irq} (Linux): time spent for servicing hardware interrupts.
\item \code{softirq} (Linux): time spent for servicing software interrupts.
\item \code{steal} (Linux 2.6.11+): time spent by other operating systems
running in a virtualized environment.
\item \code{guest} (Linux 2.6.24+): time spent running a virtual CPU for guest
operating systems under the control of the Linux kernel.
\item \code{guest_nice} (Linux 3.2.0+): time spent running a niced guest
(virtual CPU for guest operating systems under the control of the
Linux kernel).
\item \code{interrupt} (Windows): time spent for servicing hardware interrupts
( similar to “irq” on UNIX).
\item \code{dpc} (Windows): time spent servicing deferred procedure calls
(DPCs); DPCs are interrupts that run at a lower priority than
standard interrupts.
}
}
\examples{
\dontshow{if (ps::ps_is_supported() && ! ps:::is_cran_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
p <- ps_handle()
p
ps_cpu_times(p)
proc.time()
\dontshow{\}) # examplesIf}
\dontshow{if (ps::ps_is_supported()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ps_cpu_times()
\dontshow{\}) # examplesIf}
}
\seealso{
Other process handle functions: 
\code{\link{ps_children}()},
\code{\link{ps_cmdline}()},
\code{\link{ps_connections}()},
\code{\link{ps_create_time}()},
\code{\link{ps_cwd}()},
\code{\link{ps_descent}()},
\code{\link{ps_environ}()},
\code{\link{ps_exe}()},
\code{\link{ps_handle}()},
\code{\link{ps_interrupt}()},
\code{\link{ps_is_running}()},
\code{\link{ps_kill}()},
\code{\link{ps_memory_info}()},
\code{\link{ps_name}()},
\code{\link{ps_num_fds}()},
\code{\link{ps_num_threads}()},
\code{\link{ps_open_files}()},
\code{\link{ps_pid}()},
\code{\link{ps_ppid}()},
\code{\link{ps_resume}()},
\code{\link{ps_send_signal}()},
\code{\link{ps_shared_libs}()},
\code{\link{ps_status}()},
\code{\link{ps_suspend}()},
\code{\link{ps_terminal}()},
\code{\link{ps_terminate}()},
\code{\link{ps_uids}()},
\code{\link{ps_username}()}
}
\concept{process handle functions}
