% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk.R
\name{ps_fs_mount_point}
\alias{ps_fs_mount_point}
\title{Find the mount point of a file or directory}
\usage{
ps_fs_mount_point(paths)
}
\arguments{
\item{paths}{Paths to files, directories, devices, etc. They must
exist. They are normalized using \code{\link[base:normalizePath]{base::normalizePath()}}.}
}
\value{
Character vector, paths to the mount points of the input
\code{paths}.
}
\description{
Find the mount point of a file or directory
}
\examples{
\dontshow{if (ps::ps_is_supported() && ! ps:::is_cran_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ps_fs_mount_point(".")
\dontshow{\}) # examplesIf}
}
