% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill-parallel.R
\name{ps_kill_parallel}
\alias{ps_kill_parallel}
\title{Kill processes with grace period}
\usage{
ps_kill_parallel(ps, grace = 0.2)
}
\arguments{
\item{ps}{List of \href{ps_handle}{process handles}.}

\item{grace}{Grace period between \code{SIGTERM} and \code{SIGKILL} in seconds.}
}
\description{
On Unixes, \code{ps_kill_parallel()} sends a \code{SIGTERM} to all processes
in \code{ps}. If the processes are not killed after a grace period
defined by \code{grace}, a \code{SIGKILL} is issued. On Windows, the
processes are abruptly terminated without grace period.
}
\details{
Direct subprocesses are polled for termination events, which means
that \code{ps_kill_parallel()} will return earlier than the grace period
as soon as all \code{SIGTERM} signals are successful. However if \code{ps}
includes subsubprocesses, those are periodically polled for
termination as ps doesn't receive termination events for those. In
this case, \code{ps_kill_parallel()} might return a little later.
}
