% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-level.R
\name{ps_num_fds}
\alias{ps_num_fds}
\title{Number of open file descriptors}
\usage{
ps_num_fds(p)
}
\arguments{
\item{p}{Process handle.}
}
\value{
Integer scalar.
}
\description{
Note that in some IDEs, e.g. RStudio or R.app on macOS, the IDE itself
opens files from other threads, in addition to the files opened from the
main R thread.
}
\details{
For a zombie process it throws a \code{zombie_process} error.
}
\seealso{
Other process handle functions: \code{\link{ps_children}},
  \code{\link{ps_cmdline}}, \code{\link{ps_connections}},
  \code{\link{ps_cpu_times}}, \code{\link{ps_create_time}},
  \code{\link{ps_cwd}}, \code{\link{ps_environ}},
  \code{\link{ps_exe}}, \code{\link{ps_handle}},
  \code{\link{ps_interrupt}}, \code{\link{ps_is_running}},
  \code{\link{ps_kill}}, \code{\link{ps_memory_info}},
  \code{\link{ps_name}}, \code{\link{ps_num_threads}},
  \code{\link{ps_open_files}}, \code{\link{ps_pid}},
  \code{\link{ps_ppid}}, \code{\link{ps_resume}},
  \code{\link{ps_send_signal}}, \code{\link{ps_status}},
  \code{\link{ps_suspend}}, \code{\link{ps_terminal}},
  \code{\link{ps_terminate}}, \code{\link{ps_uids}},
  \code{\link{ps_username}}
}
\concept{process handle functions}
\section{Examples}{
\Sexpr[stage=install,strip.white=FALSE,results=rd]{ps:::decorate_examples('
p <- ps_handle()
ps_num_fds(p)
f <- file(tmp <- tempfile(), "w")
ps_num_fds(p)
close(f)
unlink(tmp)
ps_num_fds(p)
')}
}
