% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{ps_pid_exists}
\alias{ps_pid_exists}
\title{Check if a process id exists}
\usage{
ps_pid_exists(pid)
}
\arguments{
\item{pid}{Process id, integer scalar.}
}
\value{
Logical scalar.
}
\description{
This is faster than calling \code{\link[=ps_pids]{ps_pids()}} first.
}
