% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{ps_ppid_map}
\alias{ps_ppid_map}
\title{Parent proccesses for each process}
\usage{
ps_ppid_map()
}
\value{
A two-column data frame with integer columns \code{pid} and \code{ppid}.
}
\description{
On Windows systems the parent process might not exist any more, and
its pid might have been reused, potentially.
}
\details{
On POSIX systems, if the parent process exited already, usually PID 1
is shown instead of its id.
}
