% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-level.R
\name{ps_resume}
\alias{ps_resume}
\title{Resume (continue) a stopped process}
\usage{
ps_resume(p)
}
\arguments{
\item{p}{Process handle.}
}
\description{
Resume process execution with SIGCONT pre-emptively checking
whether PID has been reused. On Windows this has the effect of resuming
all process threads.
}
\seealso{
Other process handle functions: \code{\link{ps_children}},
  \code{\link{ps_cmdline}}, \code{\link{ps_cpu_times}},
  \code{\link{ps_create_time}}, \code{\link{ps_cwd}},
  \code{\link{ps_environ}}, \code{\link{ps_exe}},
  \code{\link{ps_handle}}, \code{\link{ps_is_running}},
  \code{\link{ps_kill}}, \code{\link{ps_memory_info}},
  \code{\link{ps_name}}, \code{\link{ps_num_threads}},
  \code{\link{ps_pid}}, \code{\link{ps_ppid}},
  \code{\link{ps_send_signal}}, \code{\link{ps_status}},
  \code{\link{ps_suspend}}, \code{\link{ps_terminal}},
  \code{\link{ps_terminate}}, \code{\link{ps_uids}},
  \code{\link{ps_username}}
}
\section{Examples}{
\Sexpr[stage=install,strip.white=FALSE,results=rd]{
ps:::decorate_examples(os = "POSIX",  '
px <- processx::process$new("sleep", "10")
p <- ps_handle(px$get_pid())
p
ps_suspend(p)
ps_status(p)
ps_resume(p)
ps_status(p)
ps_kill(p)
')}
}
