% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{ps_start}
\alias{ps_start}
\title{Start a process in the background}
\usage{
ps_start(command, args = character(), ...)
}
\arguments{
\item{command}{Character scalar, the command to run. Note that
this argument is not passed to a shell, so no tilde-expansion
or variable substitution is performed on it. It should not be
quoted with ‘base::shQuote()’. See ‘base::normalizePath()’
for tilde-expansion.}

\item{args}{Character vector, arguments to the command. They will
be used as is, without a shell. They don't need to be
escaped.}

\item{...}{Additional arguments are passed to \link[processx:process]{processx::process}.}
}
\description{
Start a process in the background with \link[processx:process]{processx::process}, and return
an object that supports both the \code{processx::process} and \code{ps::process}
methods.
}
\section{Examples}{
\preformatted{p <- ps_start("sleep", "5")
p

## processx methods
p$wait(100)
p$poll_io(100)

## ps methods
p$suspend()
p$resume()
}
}

