% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{ps_string}
\alias{ps_string}
\title{Encode a \code{ps_handle} as a short string}
\usage{
ps_string(p = ps_handle())
}
\arguments{
\item{p}{Process handle.}
}
\value{
A process string (scalar character), that can be passed to
\code{ps_handle()} in place of a pid.
}
\description{
A convenient format for passing between processes, naming semaphores, or
using as a directory/file name. Will always be 12 alphanumeric characters,
with the first character guarantied to be a letter. Encodes the pid and
creation time for a process.
}
\examples{
\dontshow{if (ps::ps_is_supported() && ! ps:::is_cran_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
(p <- ps_handle())
(str <- ps_string(p))
ps_handle(pid = str)
\dontshow{\}) # examplesIf}
}
