% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{ps_system_swap}
\alias{ps_system_swap}
\title{System swap memory statistics}
\usage{
ps_system_swap()
}
\value{
Named list. All numbers are in bytes:
\itemize{
\item \code{total}: total swap memory.
\item \code{used}: used swap memory.
\item \code{free}: free swap memory.
\item \code{percent}: the percentage usage.
\item \code{sin}: the number of bytes the system has swapped in from disk
(cumulative). This is \code{NA} on Windows.
\item \code{sout}: the number of bytes the system has swapped out from disk
(cumulative). This is \code{NA} on Windows.
}
}
\description{
System swap memory statistics
}
\examples{
\dontshow{if (ps::ps_is_supported() && ! ps:::is_cran_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ps_system_swap()
\dontshow{\}) # examplesIf}
}
\seealso{
Other memory functions: 
\code{\link{ps_system_memory}()}
}
\concept{memory functions}
