% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{ps_users}
\alias{ps_users}
\title{List users connected to the system}
\usage{
ps_users()
}
\value{
A data frame (tibble) with columns
\code{username}, \code{tty}, \code{hostname}, \code{start_time}, \code{pid}. \code{tty} and \code{pid}
are \code{NA} on Windows. \code{pid} is the process id of the login process.
For local users the \code{hostname} column is the empty string.
}
\description{
List users connected to the system
}
