% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{ps_wait}
\alias{ps_wait}
\title{Wait for some processes to finish}
\usage{
ps_wait(procs, timeout = Inf)
}
\arguments{
\item{procs}{List of \code{ps_handle} objects. Processes to wait on.}
}
\value{
TODO
}
\description{
Wait for some processes to finish
}
