% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posix.R
\name{signals}
\alias{signals}
\title{List of all supported signals}
\usage{
signals()
}
\value{
List of integers, named by signal names.
}
\description{
Only the signals supported by the current platform are included.
}
