% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{with_process}
\alias{with_process}
\title{Query process properties with caching}
\usage{
with_process(process, expr)
}
\arguments{
\item{process}{Process object to cache information for.}

\item{expr}{Expression to evaluate in the context of the cache.
Typically it contains \code{process} queries.}
}
\value{
Value of \code{expr}.
}
\description{
Avoid quering process information multiple time, as much as possible.
}
