% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{ps}
\alias{ps}
\title{Process table}
\usage{
ps(user = NULL, after = NULL)
}
\arguments{
\item{user}{Username, to filter the results to matching processes.}

\item{after}{Start time (POSIXt), to filter the results to processes
that started after this.}
}
\value{
Data frame (tibble), see columns below.

Columns:
\itemize{
\item \code{pid}: Process ID.
\item \code{ppid}: Process ID of parent process.
\item \code{name}: Process name.
\item \code{username}: Name of the user (real uid on POSIX).
\item \code{status}: I.e. \emph{running}, \emph{sleeping}, etc.
\item \code{user}: User CPU time.
\item \code{system}: System CPU time.
\item \code{rss}: Resident set size, the amount of memory the process currently
uses. Does not include memory that is swapped out. It does include
shared libraries.
\item \code{vms}: Virtual memory size. All memory the process has access to.
\item \code{created}: Time stamp when the process was created.
\item \code{ps_handle}: \code{ps_handle} objects, in a list column.
}
}
\description{
Process table
}
