% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{ps_cpu_count}
\alias{ps_cpu_count}
\title{Number of logical or physical CPUs}
\usage{
ps_cpu_count(logical = TRUE)
}
\arguments{
\item{logical}{If \code{FALSE} the number of phyical cores only. If cannot
be determined, it returns \code{NA}.}
}
\value{
Integer scalar, the number of CPUs.
}
\description{
Number of logical or physical CPUs
}
