% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ps_os_type}
\alias{ps_os_type}
\title{Query the type of the OS}
\usage{
ps_os_type()
}
\value{
\code{ps_os_type} return a named logical vector. The rest of the
functions return a logical scalar.
}
\description{
Query the type of the OS
}
