% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ps_os_type}
\alias{ps_os_type}
\alias{ps_os_is_posix}
\alias{ps_os_is_windows}
\alias{ps_os_is_linux}
\alias{ps_os_is_osx}
\alias{ps_os_is_macos}
\alias{ps_os_is_freebsd}
\alias{ps_os_is_openbsd}
\alias{ps_os_is_netbsd}
\alias{ps_os_is_bsd}
\alias{ps_os_is_sunos}
\alias{ps_os_is_aix}
\title{Query the type of the OS}
\usage{
ps_os_type()

ps_os_is_posix()

ps_os_is_windows()

ps_os_is_linux()

ps_os_is_osx()

ps_os_is_macos()

ps_os_is_freebsd()

ps_os_is_openbsd()

ps_os_is_netbsd()

ps_os_is_bsd()

ps_os_is_sunos()

ps_os_is_aix()
}
\value{
\code{ps_os_type} return a named logical vector. The rest of the
functions return a logical scalar.
}
\description{
Query the type of the OS
}
